rename keycodes.js to key-codes.js
