+++ extensions/amp-carousel/0.1/base-carousel.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {KeyCodes} from '../../../src/utils/keycodes';
+import {KeyCodes} from '../../../src/utils/key-codes';
 import {timerFor} from '../../../src/services';
 
 /**
