+++ extensions/amp-sidebar/0.1/test/test-amp-sidebar.js
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import {KeyCodes} from '../../../../src/utils/keycodes';
+import {KeyCodes} from '../../../../src/utils/key-codes';
 import {adopt} from '../../../../src/runtime';
 import {createIframePromise} from '../../../../testing/iframe';
 import {platformFor} from '../../../../src/services';
